+++ src/tokens.h
@@ -7,7 +7,7 @@ enum TOKEN {
     BEGIN=0, END, READ, WRITE, IF, THEN, ELSE, ENDIF, WHILE, ENDWHILE, ID, INTLITERAL, FALSEOP, TRUEOP, NULLOP, LPAREN, RPAREN, SEMICOLON, COMMA, ASSIGNOP, PLUSOP, MINUSOP, MULTOP,DIVOP, NOTOP, LESSOP, LESSEQUALOP, GREATEROP, GREATEREQUALOP, EQUALOP, NOTEQUALOP, SCANEOF, ERROR
 };
 
-typedef struct {
+struct TokenCatch{
     /* A type corresponding to the TOKEN enum. */
     short token;
     /* The character that was found. */
@@ -17,7 +17,7 @@ typedef struct {
     /* The column where it started. */
     int col_no;
 
-} TokenCatch;
+};
 
 /*! Token_GetName gets a character string representing a token. 
 
@@ -25,9 +25,11 @@ typedef struct {
     \return const char* A string from a lookup table, e.g, "BEGIN". If the param is not a valid token, then it returns "NULL".
     \author klm127
     \date 2/7/2023
-    \note Not Covered By Unit Tests
+    \note Covered By Unit Tests
 */
 const char * Token_GetName(int id);
+
+
 
 /*! Token_Recognize recognizes a token from a string parameter and returns the corresponding token id.
     \param word The string to identify.
@@ -35,6 +37,7 @@ const char * Token_GetName(int id);
     \pre The token is only alphanumeric characters and has already been validated not to contain a comment. (Two '--' chars.)
     \pre The token does not contain any valid operators. (It will be recognized as an identifier if so.)
     \return int a integer within the TOKEN enum; ERROR if it was not a valid word.
+    \note Covered by unit tests.
 */
 int Token_RecognizeKeyword(char * word, int length);
 
@@ -47,7 +50,14 @@ int Token_RecognizeKeyword(char * word, int length);
     \param col_found_at The column at which the token was found. 
     \returns A new TokenCatch encapsulating the parameter data. 
 */
-TokenCatch Token_Catch(short tokenType, char* raw_text_found, int line_found_at, int col_found_at);
+struct TokenCatch* Token_Catch(short tokenType, char* raw_text_found, int line_found_at, int col_found_at);
+
+/*! Token_GetOpName gets a malloced string for assignment to raw representing what must have been found for an operator text given an enumerated operarator token. If its not one of the operators, it returns ':', which is the one case when a valid operator character was a syntactic error.
+
+    \param tokenType The operator token enumerated id
+    \return A malloced string containing the operator, e.g. "<=".
+*/
+char * Token_GetOpRaw(short tokenType);
 
 /*!
     Token_Catch_Op is called when an op is found. It still produces a TokenCatch but it infers the text that was found based on the token type rather than needing the raw text, since there is not variation in how the operators can be written.  
@@ -57,7 +67,7 @@ TokenCatch Token_Catch(short tokenType, char* raw_text_found, int line_found_at,
     \param col_found_at The column at which the token was found. 
     \returns A new TokenCatch encapsulating the parameter data. 
 */
-TokenCatch Token_CatchOp(short tokenType, int line_found_at, int col_found_at);
+struct TokenCatch* Token_CatchOp(short tokenType, int line_found_at, int col_found_at);
 
 /*!
     Token_CatchError is called when an error is found. Whatever character is passed in will become the 'raw' member of a TokenCatch. 
@@ -65,8 +75,14 @@ TokenCatch Token_CatchOp(short tokenType, int line_found_at, int col_found_at);
     \param raw_text_found A char pointer to the raw text that caused this token to be identified as such.
     \param line_found_at The line in the file the token was found.
     \param col_found_at The column at which the token was found. 
-    \returns A new TokenCatch encapsulating the parameter data. 
+    \returns A pointer to a malloced TokenCatch encapsulating the parameter data. 
 */
-TokenCatch Token_CatchError(char badChar, int line_found_at, int col_found_at);
+struct TokenCatch* Token_CatchError(char badChar, int line_found_at, int col_found_at);
+
+/*!
+    Token Destroy deallocates a token by first freeing the internal 'raw' string, then deallocating the token itself. 
+    \param token A token to deallocate.
+*/
+void Token_Destroy(struct TokenCatch* token);
 
 #endif